requirements.txt
azure-ai-inference>=1.0.0b9
azure-core>=1.34.0
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
sqlalchemy>=2.0.41
werkzeug>=3.1.3
oauthlib>=3.2.2
pyjwt>=2.10.1
flask-dance>=7.1.0
openai>=1.86.0
python-dotenv>=1.0.0
Flask-Migrate>=4.0.5
psycopg2-binary>=2.9.9
sqlalchemy-cockroachdb>=2.0.3 